+++ example/test/unit_test.dart
@@ -1,0 +1,10 @@
+import 'dart:typed_data';
+import 'dart:io';
+import 'dart:async';
+
+import 'package:flutter/material.dart';
+import 'package:flutter_test/flutter_test.dart';
+
+import 'package:mic_stream/mic_stream.dart';
+
+void main() {}